[BE] feat: 메세지큐를 사용한 숙소 번역

+++ back/oasis/src/main/java/org/muhan/oasis/review/vo/in/RegistReviewRequestVo.java
@@ -19,7 +19,7 @@ public class RegistReviewRequestVo {
     @NotNull(message = "별점은 필수입니다.")
     @DecimalMin(value = "0.5", message = "별점은 최소 0.5 이상이어야 합니다.")
     @DecimalMax(value = "5.0", message = "별점은 최대 5.0 이하여야 합니다.")
-    private float rating;
+    private BigDecimal rating;
 
     @Size(max = 2000, message = "내용은 2000자 이하여야 합니다.")
     private String originalContent;
